+++ front/lib/triggers/built-in-webhooks/jira/schemas/issue_created.ts
