+++ front/lib/triggers/built-in-webhooks/zendesk/preset.ts
@@ -1,8 +1,8 @@
 import { ZendeskOAuthExtraConfig } from "@app/components/data_source/ZendeskOAuthExtraConfig";
 import { CreateWebhookZendeskConnection } from "@app/lib/triggers/built-in-webhooks/zendesk/components/CreateWebhookZendeskConnection";
 import { WebhookSourceZendeskDetails } from "@app/lib/triggers/built-in-webhooks/zendesk/components/WebhookSourceZendeskDetails";
-import { ZENDESK_WEBHOOK_EVENTS } from "@app/lib/triggers/built-in-webhooks/zendesk/zendesk_webhook_events";
-import { ZendeskWebhookService } from "@app/lib/triggers/built-in-webhooks/zendesk/zendesk_webhook_service";
+import { ZENDESK_WEBHOOK_EVENTS } from "@app/lib/triggers/built-in-webhooks/zendesk/events";
+import { ZendeskWebhookService } from "@app/lib/triggers/built-in-webhooks/zendesk/service";
 import type { PresetWebhook } from "@app/types/triggers/webhooks_source_preset";
 
 export const ZENDESK_WEBHOOK_PRESET: PresetWebhook<"zendesk"> = {
