+++ front/lib/triggers/built-in-webhooks/zendesk/schemas/organization_created.ts
