+++ front/lib/triggers/built-in-webhooks/zendesk/schemas/ticket_comment_added.ts
