+++ front/lib/triggers/built-in-webhooks/zendesk/schemas/ticket_created.ts
