+++ front/lib/triggers/built-in-webhooks/zendesk/schemas/ticket_priority_changed.ts
