+++ front/lib/triggers/built-in-webhooks/zendesk/schemas/ticket_status_changed.ts
