+++ front/lib/triggers/built-in-webhooks/zendesk/schemas/user_created.ts
